moved exceptions to single file
